requirements/base.txt
# Static Files
whitenoise>=6.6.0  # https://github.com/evansd/whitenoise
argon2-cffi>=23.1.0
argon2-cffi-bindings>=21.2.0
# Django
# ------------------------------------------------------------------------------
django>=4.2.6  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ>=0.11.2  # https://github.com/joke2k/django-environ
django-model-utils>=4.3.1  # https://github.com/jazzband/django-model-utils
# Django REST Framework
djangorestframework<=3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers>=4.3.0  # https://github.com/adamchainz/django-cors-headers
django-filter>=23.3.0
djangorestframework-jsonapi>=6.1.0
drf-nested-routers>=0.90.0
django-rest-framework-nested>=0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-extensions>=3.2.3
django-allauth>=0.57.0
django-redis>=5.4.0
# DRF-spectacular for api documentation
drf-spectacular>=0.26.5  # https://github.com/tfranzel/drf-spectacular
# Google Classroom API
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.104.0
# NLP model
torch>=2.1.0
transformers>=4.35.0
